+++ docs/changelog.md
@@ -5,6 +5,7 @@ List of changes made to waiter.
 # waiter 0.2.1.9000
 
 - Fixed `waiter_hide_on_render` see [#79](https://github.com/JohnCoene/waiter/issues/79).
+- Added `waiter_preloader` shows the full page loading screen when the app is loaded and automatically removes it when all the UI is rendered: only runs once. [#82](https://github.com/JohnCoene/waiter/issues/82)
 
 ## waiter 0.2.0